+++ internal/app/server/server_test.go
@@ -192,7 +192,7 @@ func Test_serv(t *testing.T) {
 					Message: protocol.Message{
 						Type: protocol.MessageTypeRequest,
 					},
-					Payload: []int{1, 2, 3},
+					Payload: []int64{1, 2, 3},
 				}
 				requestBuffer := &bytes.Buffer{}
 				require.NoError(t, gob.NewEncoder(requestBuffer).Encode(request))
@@ -222,7 +222,7 @@ func Test_serv(t *testing.T) {
 					Message: protocol.Message{
 						Type: protocol.MessageTypeResponse,
 					},
-					Payload: []int{1, 2, 3},
+					Payload: []int64{1, 2, 3},
 				}
 				requestBuffer := &bytes.Buffer{}
 				require.NoError(t, gob.NewEncoder(requestBuffer).Encode(request))
